SMoved and splitted scatter3d series into a subfolder.
