+++ js/masters/highcharts-3d.src.js
@@ -17,4 +17,4 @@ import '../Core/Axis/Tick3D.js';
 import '../Core/Series/Series3D.js';
 import '../Series/Column3DSeries.js';
 import '../Series/Pie3D/Pie3DComposition.js';
-import '../Series/Scatter3DSeries.js';
+import '../Series/Scatter3D/Scatter3DSeries.js';
