+++ ts/Series/Scatter3D/Scatter3DPointOptions.d.ts
@@ -1,0 +1,29 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ScatterPointOptions from '../Scatter/ScatterPointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface Scatter3DPointOptions extends ScatterPointOptions {
+    z?: number;
+}
+
+export default Scatter3DPointOptions;
