+++ ts/Series/Scatter3D/Scatter3DSeriesOptions.d.ts
@@ -1,0 +1,29 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ScatterSeriesOptions from '../Scatter/ScatterSeriesOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface Scatter3DSeriesOptions extends ScatterSeriesOptions {
+    // nothing here yet
+}
+
+export default Scatter3DSeriesOptions;
